Remove the style variation hook
