+++ client/components/themenew/theme.jsx
@@ -1,33 +1,46 @@
 import { isDefaultGlobalStylesVariationSlug, ThemeCard } from '@automattic/design-picker';
+import page from 'page';
 import ThemeTypeBadge from 'calypso/components/theme-type-badge';
 import { decodeEntities } from 'calypso/lib/formatting';
 import { useSelector } from 'calypso/state';
 import { useSiteGlobalStylesStatus } from 'calypso/state/sites/hooks/use-site-global-styles-status';
 import { getSiteSlug } from 'calypso/state/sites/selectors';
-import { isInstallingTheme, isThemeActive } from 'calypso/state/themes/selectors';
+import {
+	getThemeDetailsUrl,
+	isInstallingTheme,
+	isThemeActive,
+} from 'calypso/state/themes/selectors';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
 import useImageClick from './hooks/use-image-click';
-import useSelectedStyleVariation from './hooks/use-selected-style-variation';
+import useThemeShowcaseTracks from './hooks/use-theme-showcase-tracks';
 import useThemesBookmark from './hooks/use-themes-bookmark';
 import ThemeActions from './theme-actions';
 import { useThemeContext } from './theme-context';
 import ThemePlaceholder from './theme-placeholder';
+import { useThemeShowcaseContext } from './theme-showcase-context';
 import ThemeThumbnail from './theme-thumbnail';
 import ThemeUpdateAlert from './theme-update-alert';
 
 export default function Theme( { isPlaceholder } ) {
-	const { selectedStyleVariation, theme, themeId } = useThemeContext();
+	const { selectedStyleVariation, setSelectedStyleVariation, theme, themeId } = useThemeContext();
+	const { tabFilter } = useThemeShowcaseContext();
 
 	const siteId = useSelector( getSelectedSiteId );
 	const siteSlug = useSelector( ( state ) => getSiteSlug( state, siteId ) );
+	const themeDetailsUrl = useSelector( ( state ) =>
+		getThemeDetailsUrl( state, themeId, siteId, {
+			styleVariationSlug: selectedStyleVariation?.slug,
+			tabFilter,
+		} )
+	);
 	const isActive = useSelector( ( state ) => isThemeActive( state, themeId, siteId ) );
 	const isInstalling = useSelector( ( state ) => isInstallingTheme( state, themeId, siteId ) );
 
-	const { onStyleVariationClick, onStyleVariationMoreClick } = useSelectedStyleVariation();
 	const { imageClickUrl, imageLabel, onImageClick } = useImageClick( { selectedStyleVariation } );
 	const { shouldLimitGlobalStyles } = useSiteGlobalStylesStatus( siteId );
 
 	const { bookmarkRef, setThemesBookmark } = useThemesBookmark();
+	const { recordThemeClick } = useThemeShowcaseTracks();
 
 	if ( isPlaceholder || ! theme ) {
 		return <ThemePlaceholder />;
@@ -37,6 +50,17 @@ export default function Theme( { isPlaceholder } ) {
 
 	const isLockedStyleVariation =
 		shouldLimitGlobalStyles && ! isDefaultGlobalStylesVariationSlug( selectedStyleVariation?.slug );
+
+	const onStyleVariationClick = ( styleVariation ) => {
+		recordThemeClick( 'calypso_themeshowcase_theme_style_variation_click', {
+			styleVariationSlug: styleVariation?.slug,
+		} );
+		setSelectedStyleVariation( styleVariation );
+	};
+	const onStyleVariationMoreClick = () => {
+		recordThemeClick( 'calypso_themeshowcase_theme_style_variation_more_click' );
+		page( themeDetailsUrl );
+	};
 
 	return (
 		<ThemeCard
